--- src/app/mypage/playlist/page.module.css
@@ -1,11 +1,0 @@
-.actionButtonContainer {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  margin-top: 0.5rem;
-  margin-right: 1rem;
-}
-
-.actionButtonContainer > :first-child {
-  margin-right: 0.5rem;
-}
